Created a simple test framework according to the book Test-driven development by example, authored by kent beck.
The book wrote the tests in python, which i converted to java.

TestCase is the main function which runs the test case.
TestCaseTest runs different tests.
TestResult gathers the result of each test, whether it passes or fails.
TestSuite runs the the sampled tests.
WasRun is a simple test class