Refactored runtime, testcase, testcasetest and testruntime
In runtime, i added a new arraylist so that the references in the list
being sent was copied to the new arraylist. This way we ensure that the
sending part doesn't mix with the list.

recatored testcase. Changed name of initializeTeardown and setup with
runTeardown and runSetup. Added a finally clause to ensure that teardown
is run, in case of exception.
Also changed the one parameter constructor to send a this to the second
constructor referencing itself as a new instance.

added two test to try the new nullpointerexception in TestCase.run, to
ensure neither methodname nor instance is null.

changed from OurTest to OurFramework in our TestRuntime.