Refactored Runtime
made classList private instead of public to keep a better encapsulation.

Removed unused commented code.

Changed the first addClass to take the input list in a new list, to copy
the references.

Refactored names to be more descriptive.

Changed execute and moved relevent code to be put in correct methods.